CCP-461: reorganized files to eliminate duplcated code
