+++ frontend/components/Price/style.js
