+++ frontend/components/PriceStriked/components/PriceStriked/index.jsx
