+++ frontend/components/PriceStriked/components/PriceStriked/style.js
