+++ frontend/components/PriceStriked/style.js
