+++ frontend/pages/product/ProductDetailPrice/components/Price/style.js
